requirements.txt

viam-sdk==0.35.0

typing-extensions
